// frontend/src/common/enums/app/reducer-name.enum.ts
enum ReducerName {
  LOGIN = 'login',
  SIGN_UP = 'sign-up',
  USER = 'user',
  COUNTER = 'counter',
  PATIENT = 'patient',
  PRACTITIONER = 'practitioner',
  CONTACT = 'contact',
  MEDICAL_INFO = 'medicalInfo',
  BMI_INFO = 'bmiInfo',
  DOCUMENT_UPLOAD = 'documentUpload',
  EDUCATION = 'education',
  WORK_EXPERIENCE = 'work_experience',
  NOTIFICATION_SETTINGS = 'notification_settings',

  PATIENT_INFO = 'patient-info',
  PATIENT_BMI = 'patient-bmi',
  PATIENT_MED_INFO = 'patient-med-info',
  PATIENT_CONTACT_PEOPLE = 'patient-contact-people',
  PATIENT_NOTIFICATIONS = 'patient-notifications',

  DECLARATION = 'declaration',
  PRACTITIONER_REGISTRATION = 'practitioner_registration',
  MY_PRACTITIONER = 'my_practitioner',
  DECLARATION_FILTERS = 'declaration_filters',
  LOAD_DOCUMENTS = 'load_documents',
  APPOINTMENTS = 'appointments',
  CHAT = 'chat',
}

export { ReducerName };